move pydantic models alongside the class they are used in to remove the number of files
