+++ mobility/motives/motive.py
@@ -1,10 +1,12 @@
+from __future__ import annotations
+
 import pandas as pd
 import polars as pl
 
-from typing import List, Dict
+from typing import Annotated, List, Dict
 
 from mobility.in_memory_asset import InMemoryAsset
-from mobility.motives.motive_parameters import MotiveParameters
+from pydantic import BaseModel, ConfigDict, Field
 
 class Motive(InMemoryAsset):
 
@@ -24,7 +26,7 @@ class Motive(InMemoryAsset):
             country_utilities: Dict = None,
             sink_saturation_coeff: float = None,
             extra_inputs: dict | None = None,
-            parameters: MotiveParameters | None = None,
+            parameters: "MotiveParameters" | None = None,
         ):
 
         parameters = self.prepare_parameters(
@@ -99,3 +101,87 @@ class Motive(InMemoryAsset):
     
 
 
+
+class MotiveParameters(BaseModel):
+    """Common parameters for motives."""
+
+    model_config = ConfigDict(extra="forbid")
+
+    value_of_time: Annotated[
+        float,
+        Field(
+            default=10.0,
+            ge=0.0,
+            title="Value of time",
+            description="Utility weight for time spent traveling or on activities.",
+        ),
+    ]
+
+    saturation_fun_ref_level: Annotated[
+        float,
+        Field(
+            default=1.5,
+            ge=0.0,
+            title="Saturation reference level",
+            description="Reference level used by the sink saturation utility function.",
+        ),
+    ]
+
+    saturation_fun_beta: Annotated[
+        float,
+        Field(
+            default=4.0,
+            ge=0.0,
+            title="Saturation beta",
+            description="Shape parameter of the sink saturation utility function.",
+        ),
+    ]
+
+    value_of_time_v2: Annotated[
+        float | None,
+        Field(
+            default=None,
+            ge=0.0,
+            title="Alternative value of time",
+            description="Optional alternative value of time for second utility formulation.",
+        ),
+    ]
+
+    survey_ids: Annotated[
+        list[str] | None,
+        Field(
+            default=None,
+            title="Survey motive IDs",
+            description="List of survey-specific IDs mapped to this motive.",
+        ),
+    ]
+
+    radiation_lambda: Annotated[
+        float | None,
+        Field(
+            default=None,
+            ge=0.0,
+            le=1.0,
+            title="Radiation model lambda",
+            description="Radiation-model parameter controlling destination choice dispersion.",
+        ),
+    ]
+
+    country_utilities: Annotated[
+        dict[str, float] | None,
+        Field(
+            default=None,
+            title="Country utility offsets",
+            description="Optional country-level utility offsets used for this motive.",
+        ),
+    ]
+
+    sink_saturation_coeff: Annotated[
+        float,
+        Field(
+            default=1.0,
+            ge=0.0,
+            title="Sink saturation coefficient",
+            description="Coefficient scaling sink saturation in activity utility.",
+        ),
+    ]
