+++ mobility/parsers/mobility_survey/mobility_survey.py
@@ -1,12 +1,12 @@
+from __future__ import annotations
 import os
 import pathlib
 import polars as pl
 import pandas as pd
-
+from typing import Annotated
+from pydantic import BaseModel, ConfigDict, Field
 from mobility.file_asset import FileAsset
-from mobility.parsers.mobility_survey.mobility_survey_parameters import (
-    MobilitySurveyParameters,
-)
+
 
 class MobilitySurvey(FileAsset):
     """
@@ -25,7 +25,7 @@ class MobilitySurvey(FileAsset):
         survey_name: str | None = None,
         country: str | None = None,
         seq_prob_cutoff: float | None = None,
-        parameters: MobilitySurveyParameters | None = None,
+        parameters: "MobilitySurveyParameters" | None = None,
     ):
         """Initialize mobility survey asset inputs and cache paths.
 
@@ -420,3 +420,37 @@ class MobilitySurvey(FileAsset):
             
             
         
+class MobilitySurveyParameters(BaseModel):
+    """Parameters used to configure a mobility survey asset."""
+
+    model_config = ConfigDict(extra="forbid")
+
+    survey_name: Annotated[
+        str,
+        Field(
+            title="Survey name",
+            description="Identifier of the survey dataset folder to load.",
+        ),
+    ]
+
+    country: Annotated[
+        str,
+        Field(
+            title="Country code",
+            description="ISO-like country code used to map surveys to population inputs.",
+        ),
+    ]
+
+    seq_prob_cutoff: Annotated[
+        float,
+        Field(
+            default=0.5,
+            gt=0.0,
+            le=1.0,
+            title="Sequence probability cutoff",
+            description=(
+                "Cumulative contribution cutoff used to keep the most relevant "
+                "motive/mode sequences per population group."
+            ),
+        ),
+    ]
