+++ mobility/transport_modes/carpool/detailed/__init__.py
@@ -1,4 +1,4 @@
-from .detailed_carpool_routing_parameters import DetailedCarpoolRoutingParameters
-from .detailed_carpool_generalized_cost_parameters import DetailedCarpoolGeneralizedCostParameters
+from .detailed_carpool_travel_costs import DetailedCarpoolRoutingParameters
+from .detailed_carpool_generalized_cost import DetailedCarpoolGeneralizedCostParameters
 from .detailed_carpool_generalized_cost import DetailedCarpoolGeneralizedCost
 from .detailed_carpool_travel_costs import DetailedCarpoolTravelCosts
