+++ mobility/transport_modes/carpool/detailed/detailed_carpool_generalized_cost.py
@@ -1,7 +1,10 @@
 import pandas as pd
 import numpy as np
+from typing import Annotated
+from pydantic import BaseModel, ConfigDict, Field
 
 from mobility.in_memory_asset import InMemoryAsset
+from mobility.cost_of_time_parameters import CostOfTimeParameters
 
 class DetailedCarpoolGeneralizedCost(InMemoryAsset):
     
@@ -88,4 +91,34 @@ class DetailedCarpoolGeneralizedCost(InMemoryAsset):
         costs = pd.concat([costs, ret_costs])
         
         return costs
-            
+
+
+class DetailedCarpoolGeneralizedCostParameters(BaseModel):
+    model_config = ConfigDict(extra="forbid")
+
+    number_persons: Annotated[int, Field(default=2, ge=1)]
+
+    car_cost_of_time: Annotated[CostOfTimeParameters, Field(default_factory=CostOfTimeParameters)]
+    carpooling_cost_of_time: Annotated[CostOfTimeParameters, Field(default_factory=CostOfTimeParameters)]
+
+    cost_of_time_od_coeffs: Annotated[list[dict[str, list[str] | float]], Field(
+        default_factory=lambda: [{
+            "local_admin_unit_id_from": [],
+            "local_admin_unit_id_to": [],
+            "coeff": 1.0
+        }]
+    )]
+
+    car_cost_of_distance: Annotated[float, Field(default=0.1, ge=0.0)]
+    carpooling_cost_of_distance: Annotated[float, Field(default=0.05, ge=0.0)]
+
+    car_cost_constant: Annotated[float, Field(default=0.0)]
+    carpooling_cost_constant: Annotated[float, Field(default=0.0)]
+
+    revenue_distance_local_admin_units_ids: Annotated[list[str], Field(default_factory=list)]
+    revenue_distance_r0: Annotated[float, Field(default=1.5, ge=0.0)]
+    revenue_distance_r1: Annotated[float, Field(default=0.1, ge=0.0)]
+    revenue_distance_max: Annotated[float, Field(default=3.0, ge=0.0)]
+
+    revenue_passengers_local_admin_units_ids: Annotated[list[str], Field(default_factory=list)]
+    revenue_passengers_r1: Annotated[float, Field(default=1.5, ge=0.0)]
