+++ mobility/transport_zones.py
@@ -1,4 +1,5 @@
 """Module providing transport zones for the desired study area in France and Switzerland."""
+from __future__ import annotations
 
 import os
 import logging
@@ -6,15 +7,14 @@ import geopandas as gpd
 import pathlib
 
 from importlib import resources
-from typing import Literal, List, Union
+from typing import Annotated, Literal, List, Union
 from shapely.geometry import Point
+from pydantic import BaseModel, ConfigDict, Field, model_validator
 
 from mobility.file_asset import FileAsset
-from mobility.study_area import StudyArea
-from mobility.study_area_parameters import StudyAreaParameters
+from mobility.study_area import StudyArea, StudyAreaParameters
 from mobility.parsers.osm import OSMData
 from mobility.r_utils.r_script import RScript
-from mobility.transport_zones_parameters import TransportZonesParameters
 
 class TransportZones(FileAsset):
     """
@@ -61,7 +61,7 @@ class TransportZones(FileAsset):
             inner_radius: float | None = None,
             inner_local_admin_unit_id: List[str] | None = None,
             cutout_geometries: gpd.GeoDataFrame = None,
-            parameters: TransportZonesParameters | None = None,
+            parameters: "TransportZonesParameters" | None = None,
         ):
 
         parameters = self.prepare_parameters(
@@ -241,3 +241,81 @@ class TransportZones(FileAsset):
             transport_zones = gpd.overlay(transport_zones, cutout_geometries, how="difference")
             
         return transport_zones
+
+
+class TransportZonesParameters(BaseModel):
+
+    model_config = ConfigDict(extra="forbid")
+
+    local_admin_unit_id: Annotated[
+        Union[str, list[str]],
+        Field(
+            title="Study area local admin unit ID(s)",
+            description=(
+                "Center local admin unit ID, or a list of local admin unit IDs "
+                "to define the study area."
+            ),
+        ),
+    ]
+
+    radius: Annotated[
+        float,
+        Field(
+            default=40.0,
+            ge=5.0,
+            le=100.0,
+            title="Study area radius",
+            description="Radius in km around the selected local admin unit.",
+            json_schema_extra={"unit": "km"},
+        ),
+    ]
+
+    level_of_detail: Annotated[
+        Literal[0, 1],
+        Field(
+            default=0,
+            title="Transport zones level of detail",
+            description=(
+                "Whether local admin units will be split into subzones "
+                "(level of detail = 1), according to their building footprint density."
+            ),
+        ),
+    ]
+
+    inner_radius: Annotated[
+        float | None,
+        Field(
+            default=None,
+            title="Study area inner radius",
+            description=(
+                "Radius in km around the selected local admin unit,used to flag "
+                "as local is_inner_zone. This can be used to filter out results "
+                "from the border of the simulated study area, where the simulation "
+                "will be less reliable."
+            ),
+            json_schema_extra={"unit": "km"},
+        ),
+    ]
+
+    inner_local_admin_unit_id: Annotated[
+        list[str] | None,
+        Field(
+            default=None,
+            title="Inner local admin unit IDs",
+            description=(
+                "List of local admin unit IDs marked as inner zones. This can be "
+                "used to filter out results from the border of the simulated "
+                "study area, where the simulation will be less reliable."
+            ),
+        ),
+    ]
+
+    @model_validator(mode="after")
+    def set_derived_defaults(self) -> "TransportZonesParameters":
+        if self.inner_radius is None:
+            self.inner_radius = self.radius
+
+        if isinstance(self.local_admin_unit_id, list) and self.inner_local_admin_unit_id is None:
+            self.inner_local_admin_unit_id = self.local_admin_unit_id
+
+        return self
